add module structure to js files
