+++ js/string-length.js
@@ -1,0 +1,4 @@
+// Check string max length function
+const checkStringLength = (string, maxlength) => string.length <= maxlength;
+
+export const stringLength = checkStringLength('some string', 1);
