+++ js/text-template.js
@@ -1,0 +1,13 @@
+// Template text for comments
+export const commentTextTemplate = 'Всё отлично!\n' +
+  'В целом всё неплохо. Но не всё.\n' +
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.\n' +
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.\n' +
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.\n' +
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!';
+
+// Get random item from array function
+export const getRandomItemFromArray = (names) => names[Math.floor(Math.random() * names.length)];
+
+// Get string from text template function
+export const splitTextToStrings = (string) => (getRandomItemFromArray(string.split('\n')));
